LayoutTests/inspector/styles/styles-commit-editing-expected.txt

Tests that editing is canceled properly after incremental editing.

Text
[expanded] 
element.style  { ()
color: red;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After append:
[expanded] 
element.style  { ()
color: green;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
